pyproject.toml
[project]
name = "agents-langgraph-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=1.0.2",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
    "notebook>=7.4.7",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.0",
]
